refactor(log): implement log-streaming at the rpc level
