--- api/proto/bonk/v0/log_streaming.proto
@@ -1,25 +1,0 @@
-// Copyright © 2025 Colden Cullen
-// SPDX-License-Identifier: MIT
-
-edition = "2023";
-package bonk.v0;
-
-import "google/protobuf/struct.proto";
-import "google/protobuf/timestamp.proto";
-
-message StreamLogsRequest {
-  int64 level = 1;
-  bool add_source = 2;
-}
-
-// This is meant to mirror [slog.Record](https://pkg.go.dev/log/slog#Record)
-message StreamLogsResponse {
-  google.protobuf.Timestamp time = 1;
-  string message = 2;
-  int64 level = 3;
-  map<string, google.protobuf.Value> attrs = 4;
-}
-
-service LogStreamingService {
-  rpc StreamLogs(StreamLogsRequest) returns (stream StreamLogsResponse);
-}
