+++ go.mod
@@ -16,7 +16,6 @@ tool (
 require (
 	cuelang.org/go v0.14.1
 	github.com/ValerySidorin/shclog v0.0.1
-	github.com/delicb/slogbuffer v0.1.0
 	github.com/go-viper/mapstructure/v2 v2.4.0
 	github.com/gohugoio/hashstructure v0.5.0
 	github.com/google/uuid v1.6.0
